fix sig/slot connection
svn path=/trunk/koffice/; revision=1024481

+++ krita/plugins/tools/defaulttools/kis_tool_gradient.h
@@ -72,7 +72,7 @@ public slots:
     void slotSetAntiAliasThreshold(double);
 
 #if defined(HAVE_OPENGL) && defined(HAVE_GLEW)
-    void slotSetPreviewOpacity(int value,bool final);
+    void slotSetPreviewOpacity(qreal value,bool final);
     void slotConfigChanged();
 #endif
